Split VM
