+++ src/vm/mod.rs
@@ -24,12 +24,16 @@ use crate::value::{AsyncTask, IoChannel, IoResult, Value};
 use crate::vm::opcode::Opcode;
 
 pub mod compiler;
+mod debug_runtime;
 pub mod debugger;
 pub mod instruction_set;
 mod io_pool;
 pub mod methods;
+mod modules;
 pub mod opcode;
 pub mod optimize;
+mod scheduler;
+mod state;
 mod symbol_table;
 
 /// Represents a single call frame on the call stack.
@@ -185,1683 +189,6 @@ pub struct VM<'a> {
 }
 
 impl<'a> VM<'a> {
-    pub fn new(source_ref: SourceRef<'a>, is: InstructionSet) -> Self {
-        let global_names = is.globals.get_all_names();
-        // Initialize hot-spot detector with loop and function metadata from the bytecode
-        let mut hotspot_detector = HotSpotDetector::new();
-
-        // Register all loops detected during compilation
-        for loop_meta in &is.loops {
-            hotspot_detector.register_loop(
-                loop_meta.header_ip,
-                loop_meta.back_edge_ip,
-                loop_meta.exit_ip,
-            );
-        }
-
-        // Register all functions detected during compilation
-        for func_meta in &is.functions {
-            hotspot_detector.register_function(&func_meta.name, func_meta.start_ip);
-        }
-
-        // Create the initial call frame for the main program
-        let main_frame = CallFrame {
-            return_ip: 0,
-            frame_pointer: 0,
-            stack_pointer: 0,
-            instructions: Rc::new(is),
-            function_name: "<main>".to_string(),
-            return_override: None,
-            module_binding: None,
-            awaiting_task: None,
-        };
-
-        let (io_wakeup_tx, io_wakeup_rx) = mpsc::channel();
-
-        Self {
-            stack: Vec::new(),
-            locals: Vec::new(),
-            call_stack: vec![main_frame],
-            exception_handlers: Vec::new(),
-            ip: 0,
-            gc_poll_counter: 0,
-            // Pre-size globals to declared symbol count so sparse/forward writes are safe.
-            globals: vec![Value::Void; global_names.len()],
-            global_names,
-            async_task_queue: VecDeque::new(),
-            suspended_main: None,
-            suspended_tasks: FxHashMap::default(),
-            task_waiters: FxHashMap::default(),
-            source_ref,
-            io_wakeup_tx,
-            io_wakeup_rx,
-            user_channels: Vec::new(),
-            debugger: None,
-            debug_mode: false,
-            // Profiling is opt-in (enabled by Program when requested)
-            hotspot_detector,
-            type_profile: TypeProfile::new(),
-            profiling_enabled: false,
-            // JIT compiler (Phase 2)
-            #[cfg(feature = "jit")]
-            jit_compiler: crate::jit::JitCompiler::new().ok(),
-            #[cfg(feature = "jit")]
-            jit_enabled: false,
-        }
-    }
-
-    pub fn new_with_module_binding(
-        source_ref: SourceRef<'a>,
-        is: InstructionSet,
-        module_key: DictKey,
-    ) -> WalrusResult<Self> {
-        let mut vm = Self::new(source_ref, is);
-        let binding = vm.build_module_binding(module_key)?;
-        vm.call_stack
-            .last_mut()
-            .expect("main frame should exist")
-            .module_binding = Some(binding);
-        Ok(vm)
-    }
-
-    /// Enable the debugger
-    pub fn enable_debugger(&mut self) {
-        self.debug_mode = true;
-        self.debugger = Some(debugger::Debugger::new());
-    }
-
-    /// Create a VM with profiling disabled (for benchmarking baseline)
-    pub fn new_without_profiling(source_ref: SourceRef<'a>, is: InstructionSet) -> Self {
-        let mut vm = Self::new(source_ref, is);
-        vm.profiling_enabled = false;
-        vm
-    }
-
-    /// Enable or disable JIT profiling
-    pub fn set_profiling(&mut self, enabled: bool) {
-        self.profiling_enabled = enabled;
-    }
-
-    /// Enable or disable JIT compilation and execution
-    #[cfg(feature = "jit")]
-    pub fn set_jit_enabled(&mut self, enabled: bool) {
-        self.jit_enabled = enabled;
-    }
-
-    /// Get JIT compilation statistics
-    #[cfg(feature = "jit")]
-    pub fn jit_stats(&self) -> Option<crate::jit::JitStats> {
-        self.jit_compiler.as_ref().map(|jit| jit.stats())
-    }
-
-    /// Get hot-spot detection statistics
-    pub fn hotspot_stats(&self) -> crate::jit::hotspot::HotSpotStats {
-        self.hotspot_detector.stats()
-    }
-
-    /// Get the type profile for analysis
-    pub fn type_profile(&self) -> &TypeProfile {
-        &self.type_profile
-    }
-
-    /// Get the current call frame (the one at the top of the call stack)
-    #[inline(always)]
-    fn current_frame(&self) -> &CallFrame {
-        // SAFETY: call_stack is never empty during VM execution (always has at least main frame)
-        unsafe { self.call_stack.last().unwrap_unchecked() }
-    }
-
-    #[inline(always)]
-    fn current_frame_index(&self) -> usize {
-        self.call_stack.len() - 1
-    }
-
-    /// Get the current frame pointer (start of current frame's locals)
-    #[inline(always)]
-    fn frame_pointer(&self) -> usize {
-        self.current_frame().frame_pointer
-    }
-
-    /// Get the current function name
-    #[inline(always)]
-    fn function_name(&self) -> &str {
-        let name = &self.current_frame().function_name;
-        if name.is_empty() { "<fn>" } else { name }
-    }
-
-    fn take_context(&mut self) -> ExecutionContext {
-        ExecutionContext {
-            stack: std::mem::take(&mut self.stack),
-            locals: std::mem::take(&mut self.locals),
-            call_stack: std::mem::take(&mut self.call_stack),
-            exception_handlers: std::mem::take(&mut self.exception_handlers),
-            ip: self.ip,
-        }
-    }
-
-    fn restore_context(&mut self, context: ExecutionContext) {
-        self.stack = context.stack;
-        self.locals = context.locals;
-        self.call_stack = context.call_stack;
-        self.exception_handlers = context.exception_handlers;
-        self.ip = context.ip;
-    }
-
-    /// Helper to access heap - uses thread-local ARENA
-    ///
-    /// # Safety
-    /// Uses unsafe get_arena_ptr() for performance. This is safe because:
-    /// - The VM is single-threaded
-    /// - We only access the arena within VM methods (no escaping references)
-    /// - No concurrent borrows occur within VM execution
-    #[inline]
-    pub(crate) fn get_heap(&self) -> &crate::arenas::ValueHolder {
-        unsafe { &*crate::arenas::get_arena_ptr() }
-    }
-
-    /// Helper to access heap mutably - uses thread-local ARENA
-    ///
-    /// # Safety
-    /// See get_heap() for safety rationale.
-    #[inline]
-    pub(crate) fn get_heap_mut(&mut self) -> &mut crate::arenas::ValueHolder {
-        unsafe { &mut *crate::arenas::get_arena_ptr() }
-    }
-
-    #[inline(always)]
-    pub(crate) fn source_ref(&self) -> SourceRef<'a> {
-        self.source_ref
-    }
-
-    #[inline(always)]
-    fn current_module_binding(&self) -> Option<Rc<VmModuleBinding>> {
-        self.call_stack
-            .last()
-            .and_then(|frame| frame.module_binding.clone())
-    }
-
-    fn build_module_binding(&mut self, module_key: DictKey) -> WalrusResult<Rc<VmModuleBinding>> {
-        let global_names = self.global_names.clone();
-        let mut global_values = Vec::with_capacity(global_names.len());
-
-        for name in &global_names {
-            let key = Value::String(self.get_heap_mut().push_ident(name));
-            let value = self
-                .get_heap()
-                .get_module(module_key)?
-                .get(&key)
-                .copied()
-                .unwrap_or(Value::Void);
-            global_values.push(value);
-        }
-
-        Ok(Rc::new(VmModuleBinding {
-            module_key,
-            global_names: Rc::new(global_names),
-            global_values: Rc::new(global_values),
-            source: Rc::new(self.source_ref.source().to_string()),
-            filename: Rc::new(self.source_ref.filename().to_string()),
-        }))
-    }
-
-    fn undefined_global_error(
-        &self,
-        index: usize,
-        span: Span,
-        binding: Option<&VmModuleBinding>,
-    ) -> WalrusError {
-        let name = binding
-            .and_then(|ctx| ctx.global_names.get(index))
-            .or_else(|| self.global_names.get(index))
-            .cloned()
-            .unwrap_or_else(|| format!("<global[{index}]>"));
-
-        let (src, filename) = if let Some(ctx) = binding {
-            (ctx.source.to_string(), ctx.filename.to_string())
-        } else {
-            (
-                self.source_ref.source().to_string(),
-                self.source_ref.filename().to_string(),
-            )
-        };
-
-        WalrusError::UndefinedVariable {
-            name,
-            span,
-            src,
-            filename,
-        }
-    }
-
-    fn load_global_value(&mut self, index: usize, span: Span) -> WalrusResult<Value> {
-        if let Some(binding) = self.current_module_binding() {
-            let Some(name) = binding.global_names.get(index) else {
-                return Err(self.undefined_global_error(index, span, Some(binding.as_ref())));
-            };
-
-            let key = Value::String(self.get_heap_mut().push_ident(name));
-            let module = self.get_heap().get_module(binding.module_key)?;
-            if let Some(value) = module.get(&key).copied() {
-                return Ok(value);
-            }
-
-            Ok(binding
-                .global_values
-                .get(index)
-                .copied()
-                .unwrap_or(Value::Void))
-        } else {
-            self.globals
-                .get(index)
-                .copied()
-                .ok_or_else(|| self.undefined_global_error(index, span, None))
-        }
-    }
-
-    fn store_global_value(&mut self, index: usize, value: Value, span: Span) -> WalrusResult<()> {
-        if let Some(binding) = self.current_module_binding() {
-            let Some(name) = binding.global_names.get(index) else {
-                return Err(self.undefined_global_error(index, span, Some(binding.as_ref())));
-            };
-
-            let key = Value::String(self.get_heap_mut().push_ident(name));
-            let value = self.bind_exported_value_to_module(value, &binding)?;
-            self.get_heap_mut()
-                .get_mut_module(binding.module_key)?
-                .insert(key, value);
-            Ok(())
-        } else {
-            if index >= self.globals.len() {
-                self.globals.resize(index + 1, Value::Void);
-            }
-            self.globals[index] = value;
-            Ok(())
-        }
-    }
-
-    fn bind_exported_value_to_module(
-        &mut self,
-        value: Value,
-        binding: &Rc<VmModuleBinding>,
-    ) -> WalrusResult<Value> {
-        match value {
-            Value::Function(func_key) => {
-                let function = self.get_heap().get_function(func_key)?.clone();
-                match function {
-                    WalrusFunction::Vm(mut vm_func) => {
-                        vm_func.module_binding = Some(binding.clone());
-                        Ok(self
-                            .get_heap_mut()
-                            .push(HeapValue::Function(WalrusFunction::Vm(vm_func))))
-                    }
-                    _ => Ok(value),
-                }
-            }
-            Value::StructDef(struct_key) => {
-                let original = self.get_heap().get_struct_def(struct_key)?.clone();
-                let mut rebound =
-                    crate::structs::StructDefinition::new(original.name().to_string());
-
-                for (method_name, method_fn) in original.methods() {
-                    let bound_method = match method_fn.clone() {
-                        WalrusFunction::Vm(mut vm_func) => {
-                            vm_func.module_binding = Some(binding.clone());
-                            WalrusFunction::Vm(vm_func)
-                        }
-                        other => other,
-                    };
-                    rebound.add_method(method_name.clone(), bound_method);
-                }
-
-                Ok(self.get_heap_mut().push(HeapValue::StructDef(rebound)))
-            }
-            _ => Ok(value),
-        }
-    }
-
-    pub fn export_globals_as_module(&mut self) -> WalrusResult<Value> {
-        let global_names = self.global_names.clone();
-        let global_values = self.globals.clone();
-        let mut exports = FxHashMap::default();
-
-        for (index, name) in global_names.iter().enumerate() {
-            if name == "_" {
-                continue;
-            }
-
-            let value = match self.globals.get(index).copied() {
-                Some(value) => value,
-                None => continue,
-            };
-
-            let key = self.get_heap_mut().push(HeapValue::String(name));
-            exports.insert(key, value);
-        }
-
-        let module_key = match self.get_heap_mut().push(HeapValue::Module(exports)) {
-            Value::Module(key) => key,
-            _ => unreachable!("module export must allocate a module"),
-        };
-
-        let binding = Rc::new(VmModuleBinding {
-            module_key,
-            global_names: Rc::new(global_names),
-            global_values: Rc::new(global_values),
-            source: Rc::new(self.source_ref.source().to_string()),
-            filename: Rc::new(self.source_ref.filename().to_string()),
-        });
-
-        // Rebind exported values that contain VM functions so global accesses
-        // resolve against this module, not the importing VM's global vector.
-        let names = binding.global_names.clone();
-        for name in names.iter() {
-            if name == "_" {
-                continue;
-            }
-
-            let entry_value = {
-                let key = Value::String(self.get_heap_mut().push_ident(name));
-                let module = self.get_heap().get_module(module_key)?;
-                module.get(&key).copied()
-            };
-
-            let Some(entry_value) = entry_value else {
-                continue;
-            };
-
-            let bound_value = self.bind_exported_value_to_module(entry_value, &binding)?;
-
-            let key = Value::String(self.get_heap_mut().push_ident(name));
-            self.get_heap_mut()
-                .get_mut_module(module_key)?
-                .insert(key, bound_value);
-        }
-
-        Ok(Value::Module(module_key))
-    }
-
-    /// Collect all root values that the GC needs to trace from
-    pub(crate) fn collect_roots(&self) -> Vec<Value> {
-        fn extend_context_roots(
-            roots: &mut Vec<Value>,
-            stack: &[Value],
-            locals: &[Value],
-            call_stack: &[CallFrame],
-        ) {
-            roots.extend(stack.iter().copied());
-            roots.extend(locals.iter().copied());
-            for frame in call_stack {
-                roots.extend(frame.instructions.constants.iter().copied());
-                if let Some(binding) = &frame.module_binding {
-                    roots.push(Value::Module(binding.module_key));
-                }
-                if let Some(task_key) = frame.awaiting_task {
-                    roots.push(Value::Task(task_key));
-                }
-            }
-        }
-
-        let mut roots = Vec::with_capacity(
-            self.stack.len()
-                + self.locals.len()
-                + self.globals.len()
-                + self.async_task_queue.len()
-                + self.suspended_tasks.len() * 16
-                + 128,
-        );
-
-        extend_context_roots(&mut roots, &self.stack, &self.locals, &self.call_stack);
-
-        roots.extend(self.globals.iter().copied());
-
-        for &task_key in &self.async_task_queue {
-            roots.push(Value::Task(task_key));
-        }
-
-        if let Some(suspended) = &self.suspended_main {
-            extend_context_roots(
-                &mut roots,
-                &suspended.context.stack,
-                &suspended.context.locals,
-                &suspended.context.call_stack,
-            );
-            roots.push(Value::Task(suspended.waiting_on));
-        }
-
-        for (&task_key, suspended) in &self.suspended_tasks {
-            roots.push(Value::Task(task_key));
-            roots.push(Value::Task(suspended.waiting_on));
-            extend_context_roots(
-                &mut roots,
-                &suspended.context.stack,
-                &suspended.context.locals,
-                &suspended.context.call_stack,
-            );
-        }
-
-        roots.extend(crate::program::cached_module_roots());
-
-        roots
-    }
-
-    /// Run garbage collection if needed
-    fn maybe_collect_garbage(&mut self) {
-        if self.get_heap().should_collect() {
-            let roots = self.collect_roots();
-            let freed = self.get_heap_mut().collect_garbage(&roots);
-            if freed > 0 {
-                debug!("GC: Freed {} objects", freed);
-            }
-        }
-    }
-
-    /// Stringify a value using the global heap
-    fn stringify_value(&self, value: Value) -> WalrusResult<String> {
-        self.get_heap().stringify(value)
-    }
-
-    /// Format the current call stack as a human-readable stack trace
-    /// Truncates the middle if there are too many frames (like Python does)
-    fn format_stack_trace(&self) -> String {
-        if self.call_stack.len() <= 1 {
-            return String::new();
-        }
-
-        const MAX_FRAMES_TOP: usize = 5; // Show first N frames (oldest)
-        const MAX_FRAMES_BOTTOM: usize = 5; // Show last N frames (most recent)
-
-        let mut trace = String::from("\nStack trace (most recent call last):\n");
-        let len = self.call_stack.len();
-
-        if len <= MAX_FRAMES_TOP + MAX_FRAMES_BOTTOM {
-            // Show all frames
-            for (i, frame) in self.call_stack.iter().enumerate() {
-                let name = if frame.function_name.is_empty() {
-                    "<fn>"
-                } else {
-                    frame.function_name.as_str()
-                };
-                trace.push_str(&format!("  {}: {}\n", i, name));
-            }
-        } else {
-            // Show first N frames
-            for (i, frame) in self.call_stack.iter().take(MAX_FRAMES_TOP).enumerate() {
-                let name = if frame.function_name.is_empty() {
-                    "<fn>"
-                } else {
-                    frame.function_name.as_str()
-                };
-                trace.push_str(&format!("  {}: {}\n", i, name));
-            }
-
-            // Show truncation message
-            let hidden = len - MAX_FRAMES_TOP - MAX_FRAMES_BOTTOM;
-            trace.push_str(&format!("  ... {} more frames ...\n", hidden));
-
-            // Show last N frames
-            for (i, frame) in self
-                .call_stack
-                .iter()
-                .skip(len - MAX_FRAMES_BOTTOM)
-                .enumerate()
-            {
-                let actual_index = len - MAX_FRAMES_BOTTOM + i;
-                let name = if frame.function_name.is_empty() {
-                    "<fn>"
-                } else {
-                    frame.function_name.as_str()
-                };
-                trace.push_str(&format!("  {}: {}\n", actual_index, name));
-            }
-        }
-        trace
-    }
-
-    /// Call a native stdlib function
-    fn call_native(
-        &mut self,
-        native_fn: crate::function::NativeFunction,
-        args: Vec<Value>,
-        span: Span,
-    ) -> WalrusResult<Value> {
-        crate::native_registry::dispatch_native(self, native_fn, args, span)
-    }
-
-    fn create_task_for_function_key(&mut self, function_key: FuncKey, args: Vec<Value>) -> Value {
-        let task = self
-            .get_heap_mut()
-            .push(HeapValue::Task(AsyncTask::Pending {
-                function: function_key,
-                args,
-            }));
-        if let Value::Task(task_key) = task {
-            self.enqueue_task(task_key);
-        }
-        task
-    }
-
-    fn create_task_for_cloned_function(
-        &mut self,
-        function: WalrusFunction,
-        args: Vec<Value>,
-    ) -> Value {
-        let func_value = self.get_heap_mut().push(HeapValue::Function(function));
-        let Value::Function(function_key) = func_value else {
-            unreachable!("HeapValue::Function must allocate a function key")
-        };
-        self.create_task_for_function_key(function_key, args)
-    }
-
-    fn create_non_runnable_task(&mut self, task: AsyncTask) -> Value {
-        self.get_heap_mut().push(HeapValue::Task(task))
-    }
-
-    pub(crate) fn spawn_task_from_callable(
-        &mut self,
-        callable: Value,
-        args: Vec<Value>,
-        span: Span,
-    ) -> WalrusResult<Value> {
-        match callable {
-            Value::Task(task_key) => {
-                if !args.is_empty() {
-                    return Err(WalrusError::InvalidArgCount {
-                        name: "spawn".to_string(),
-                        expected: 0,
-                        got: args.len(),
-                        span,
-                        src: self.source_ref.source().into(),
-                        filename: self.source_ref.filename().into(),
-                    });
-                }
-                Ok(Value::Task(task_key))
-            }
-            Value::Function(function_key) => {
-                let function = self.get_heap().get_function(function_key)?;
-                let expected = match function {
-                    WalrusFunction::Vm(func) => func.arity,
-                    WalrusFunction::Native(func) => func.arity(),
-                };
-                if args.len() != expected {
-                    return Err(WalrusError::InvalidArgCount {
-                        name: function.to_string(),
-                        expected,
-                        got: args.len(),
-                        span,
-                        src: self.source_ref.source().into(),
-                        filename: self.source_ref.filename().into(),
-                    });
-                }
-                Ok(self.create_task_for_function_key(function_key, args))
-            }
-            other => Err(WalrusError::TypeMismatch {
-                expected: "function or task".to_string(),
-                found: other.get_type().to_string(),
-                span,
-                src: self.source_ref.source().into(),
-                filename: self.source_ref.filename().into(),
-            }),
-        }
-    }
-
-    pub(crate) fn create_sleep_task(&mut self, delay_ms: u64) -> Value {
-        let wake_at = Instant::now() + Duration::from_millis(delay_ms);
-        self.create_non_runnable_task(AsyncTask::Sleep { wake_at })
-    }
-
-    pub(crate) fn create_timeout_task(
-        &mut self,
-        task_key: crate::arenas::TaskKey,
-        timeout_ms: u64,
-    ) -> Value {
-        let deadline = Instant::now() + Duration::from_millis(timeout_ms);
-        self.create_non_runnable_task(AsyncTask::Timeout {
-            task: task_key,
-            deadline,
-        })
-    }
-
-    /// Spawn a background I/O operation on the bounded blocking I/O pool.
-    /// Returns a Task value backed by a Channel that resolves when the worker completes.
-    pub(crate) fn spawn_io<F>(&mut self, work: F) -> Value
-    where
-        F: FnOnce() -> Result<IoResult, String> + Send + 'static,
-    {
-        let rx = io_pool::submit_io(work, self.io_wakeup_tx.clone());
-        self.create_non_runnable_task(AsyncTask::Channel(IoChannel::new(rx)))
-    }
-
-    /// Convert an IoResult from a worker thread into a Value on the VM thread.
-    fn io_result_to_value(&mut self, result: IoResult) -> WalrusResult<Value> {
-        match result {
-            IoResult::Stream(stream) => {
-                let handle = crate::stdlib::store_tcp_stream(stream);
-                Ok(Value::Int(handle))
-            }
-            IoResult::Listener(listener) => {
-                let handle = crate::stdlib::store_tcp_listener(listener);
-                Ok(Value::Int(handle))
-            }
-            IoResult::Bytes(bytes) => {
-                let text = String::from_utf8(bytes).map_err(|e| WalrusError::GenericError {
-                    message: format!("I/O result contains non-UTF8 data: {e}"),
-                })?;
-                Ok(self.get_heap_mut().push(HeapValue::String(&text)))
-            }
-            IoResult::ByteCount(n) => Ok(Value::Int(n as i64)),
-            IoResult::HttpOutcome(outcome) => self.http_outcome_to_value(outcome),
-            IoResult::Void => Ok(Value::Void),
-        }
-    }
-
-    /// Convert an IoHttpOutcome into a Value (dict or Void).
-    fn http_outcome_to_value(
-        &mut self,
-        outcome: crate::value::IoHttpOutcome,
-    ) -> WalrusResult<Value> {
-        use crate::value::IoHttpOutcome;
-
-        match outcome {
-            IoHttpOutcome::Eof => Ok(Value::Void),
-            IoHttpOutcome::BadRequest(message) => {
-                let mut dict = FxHashMap::default();
-                let ok_key = self.get_heap_mut().push(HeapValue::String("ok"));
-                dict.insert(ok_key, Value::Bool(false));
-                let err_key = self.get_heap_mut().push(HeapValue::String("error"));
-                let err_val = self.get_heap_mut().push(HeapValue::String(&message));
-                dict.insert(err_key, err_val);
-                Ok(self.get_heap_mut().push(HeapValue::Dict(dict)))
-            }
-            IoHttpOutcome::Request(req) => {
-                let mut headers = FxHashMap::default();
-                let mut header_pairs = Vec::with_capacity(req.headers.len());
-                for (name, value) in &req.headers {
-                    let key = self.get_heap_mut().push(HeapValue::String(name));
-                    let val = self.get_heap_mut().push(HeapValue::String(value));
-                    headers.insert(key, val);
-                    let pair = self.get_heap_mut().push(HeapValue::List(vec![key, val]));
-                    header_pairs.push(pair);
-                }
-                let headers_value = self.get_heap_mut().push(HeapValue::Dict(headers));
-                let header_pairs_value = self.get_heap_mut().push(HeapValue::List(header_pairs));
-
-                let query_pairs = crate::stdlib::http_parse_query(&req.query);
-                let mut query = FxHashMap::default();
-                let mut query_pairs_values = Vec::with_capacity(query_pairs.len());
-                for (name, value) in query_pairs {
-                    let key = self.get_heap_mut().push(HeapValue::String(&name));
-                    let val = self.get_heap_mut().push(HeapValue::String(&value));
-                    query.insert(key, val);
-                    let pair = self.get_heap_mut().push(HeapValue::List(vec![key, val]));
-                    query_pairs_values.push(pair);
-                }
-                let query_value = self.get_heap_mut().push(HeapValue::Dict(query));
-                let query_pairs_value = self
-                    .get_heap_mut()
-                    .push(HeapValue::List(query_pairs_values));
-
-                let mut dict = FxHashMap::default();
-                let ok_key = self.get_heap_mut().push(HeapValue::String("ok"));
-                dict.insert(ok_key, Value::Bool(true));
-                let method_key = self.get_heap_mut().push(HeapValue::String("method"));
-                let method_val = self.get_heap_mut().push(HeapValue::String(&req.method));
-                dict.insert(method_key, method_val);
-                let target_key = self.get_heap_mut().push(HeapValue::String("target"));
-                let target_val = self.get_heap_mut().push(HeapValue::String(&req.target));
-                dict.insert(target_key, target_val);
-                let path_key = self.get_heap_mut().push(HeapValue::String("path"));
-                let path_val = self.get_heap_mut().push(HeapValue::String(&req.path));
-                dict.insert(path_key, path_val);
-                let query_key = self.get_heap_mut().push(HeapValue::String("query"));
-                let query_text_val = self.get_heap_mut().push(HeapValue::String(&req.query));
-                dict.insert(query_key, query_text_val);
-                let qp_key = self.get_heap_mut().push(HeapValue::String("query_params"));
-                dict.insert(qp_key, query_value);
-                let qpp_key = self.get_heap_mut().push(HeapValue::String("query_pairs"));
-                dict.insert(qpp_key, query_pairs_value);
-                let ver_key = self.get_heap_mut().push(HeapValue::String("version"));
-                let ver_val = self.get_heap_mut().push(HeapValue::String(&req.version));
-                dict.insert(ver_key, ver_val);
-                let hdr_key = self.get_heap_mut().push(HeapValue::String("headers"));
-                dict.insert(hdr_key, headers_value);
-                let hdr_pairs_key = self.get_heap_mut().push(HeapValue::String("header_pairs"));
-                dict.insert(hdr_pairs_key, header_pairs_value);
-                let body_key = self.get_heap_mut().push(HeapValue::String("body"));
-                let body_val = self.get_heap_mut().push(HeapValue::String(&req.body));
-                dict.insert(body_key, body_val);
-                let cl_key = self
-                    .get_heap_mut()
-                    .push(HeapValue::String("content_length"));
-                dict.insert(cl_key, Value::Int(req.content_length));
-
-                Ok(self.get_heap_mut().push(HeapValue::Dict(dict)))
-            }
-        }
-    }
-
-    /// Check if a task tree has any pending I/O (Channel) tasks.
-    fn task_has_pending_io(
-        &self,
-        task_key: crate::arenas::TaskKey,
-        visited: &mut FxHashSet<crate::arenas::TaskKey>,
-    ) -> WalrusResult<bool> {
-        if !visited.insert(task_key) {
-            return Ok(false);
-        }
-        let task = self.get_heap().get_task(task_key)?;
-        match task {
-            AsyncTask::Channel(_) => Ok(true),
-            AsyncTask::UserRecv { .. } => Ok(false),
-            AsyncTask::Timeout { task, .. } => self.task_has_pending_io(*task, visited),
-            AsyncTask::Gather { tasks }
-            | AsyncTask::Race { tasks }
-            | AsyncTask::AllSettled { tasks } => {
-                for &child in tasks {
-                    if self.task_has_pending_io(child, visited)? {
-                        return Ok(true);
-                    }
-                }
-                Ok(false)
-            }
-            _ => Ok(false),
-        }
-    }
-
-    pub(crate) fn create_gather_task(&mut self, tasks: Vec<crate::arenas::TaskKey>) -> Value {
-        self.create_non_runnable_task(AsyncTask::Gather { tasks })
-    }
-
-    pub(crate) fn create_race_task(&mut self, tasks: Vec<crate::arenas::TaskKey>) -> Value {
-        self.create_non_runnable_task(AsyncTask::Race { tasks })
-    }
-
-    pub(crate) fn create_all_settled_task(&mut self, tasks: Vec<crate::arenas::TaskKey>) -> Value {
-        self.create_non_runnable_task(AsyncTask::AllSettled { tasks })
-    }
-
-    pub(crate) fn create_user_channel(&mut self) -> (Value, Value) {
-        let id = self.user_channels.len();
-        self.user_channels.push(UserChannel {
-            buffer: Rc::new(RefCell::new(VecDeque::new())),
-            closed: Rc::new(RefCell::new(false)),
-        });
-        // Sender and receiver are dicts with a magic __channel_id field
-        let type_key = self.get_heap_mut().push(HeapValue::String("__type"));
-        let id_key = self.get_heap_mut().push(HeapValue::String("__channel_id"));
-        let id_val = Value::Int(id as i64);
-
-        let sender_type = self.get_heap_mut().push(HeapValue::String("sender"));
-        let mut sender_dict = FxHashMap::default();
-        sender_dict.insert(type_key, sender_type);
-        sender_dict.insert(id_key, id_val);
-        let sender = self.get_heap_mut().push(HeapValue::Dict(sender_dict));
-
-        let receiver_type = self.get_heap_mut().push(HeapValue::String("receiver"));
-        let mut receiver_dict = FxHashMap::default();
-        receiver_dict.insert(type_key, receiver_type);
-        receiver_dict.insert(id_key, id_val);
-        let receiver = self.get_heap_mut().push(HeapValue::Dict(receiver_dict));
-
-        (sender, receiver)
-    }
-
-    pub(crate) fn channel_send(
-        &mut self,
-        sender_key: crate::arenas::DictKey,
-        value: Value,
-    ) -> WalrusResult<bool> {
-        let dict = self.get_heap().get_dict(sender_key)?.clone();
-        let id_key_str = self.get_heap_mut().push(HeapValue::String("__channel_id"));
-        if let Some(Value::Int(id)) = dict.get(&id_key_str) {
-            let id = *id as usize;
-            if id < self.user_channels.len() {
-                if *self.user_channels[id].closed.borrow() {
-                    return Ok(false);
-                }
-                self.user_channels[id].buffer.borrow_mut().push_back(value);
-                return Ok(true);
-            }
-        }
-        Ok(false)
-    }
-
-    pub(crate) fn channel_recv(
-        &mut self,
-        receiver_key: crate::arenas::DictKey,
-    ) -> WalrusResult<Value> {
-        let dict = self.get_heap().get_dict(receiver_key)?.clone();
-        let id_key_str = self.get_heap_mut().push(HeapValue::String("__channel_id"));
-        if let Some(Value::Int(id)) = dict.get(&id_key_str) {
-            let id = *id as usize;
-            if id < self.user_channels.len() {
-                // Check buffer first — if data is already available, return Ready task
-                let buffered = self.user_channels[id].buffer.borrow_mut().pop_front();
-                if let Some(value) = buffered {
-                    return Ok(self.create_non_runnable_task(AsyncTask::Ready(value)));
-                }
-                // No data yet — create a pending recv task
-                return Ok(self.create_non_runnable_task(AsyncTask::UserRecv { channel_id: id }));
-            }
-        }
-        Err(WalrusError::GenericError {
-            message: "asyncx.recv: invalid receiver".to_string(),
-        })
-    }
-
-    pub(crate) fn channel_close(
-        &mut self,
-        endpoint_key: crate::arenas::DictKey,
-    ) -> WalrusResult<bool> {
-        let dict = self.get_heap().get_dict(endpoint_key)?.clone();
-        let id_key_str = self.get_heap_mut().push(HeapValue::String("__channel_id"));
-        if let Some(Value::Int(id)) = dict.get(&id_key_str) {
-            let id = *id as usize;
-            if id < self.user_channels.len() {
-                *self.user_channels[id].closed.borrow_mut() = true;
-                return Ok(true);
-            }
-        }
-        Ok(false)
-    }
-
-    fn cancelled_task_error_value(&mut self) -> Value {
-        self.get_heap_mut()
-            .push(HeapValue::String("task cancelled"))
-    }
-
-    fn cancel_task_recursive_internal(
-        &mut self,
-        task_key: crate::arenas::TaskKey,
-        visited: &mut FxHashSet<crate::arenas::TaskKey>,
-    ) -> WalrusResult<bool> {
-        if !visited.insert(task_key) {
-            return Ok(false);
-        }
-
-        let task = self.get_heap().get_task(task_key)?.clone();
-        match task {
-            AsyncTask::Pending { .. }
-            | AsyncTask::Sleep { .. }
-            | AsyncTask::Channel(_)
-            | AsyncTask::UserRecv { .. } => {
-                *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Cancelled;
-                self.suspended_tasks.remove(&task_key);
-                Ok(true)
-            }
-            AsyncTask::Timeout { task, .. } => {
-                let _ = self.cancel_task_recursive_internal(task, visited)?;
-                *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Cancelled;
-                self.suspended_tasks.remove(&task_key);
-                Ok(true)
-            }
-            AsyncTask::Gather { tasks }
-            | AsyncTask::Race { tasks }
-            | AsyncTask::AllSettled { tasks } => {
-                for child in tasks {
-                    let _ = self.cancel_task_recursive_internal(child, visited)?;
-                }
-                *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Cancelled;
-                self.suspended_tasks.remove(&task_key);
-                Ok(true)
-            }
-            AsyncTask::Ready(_) | AsyncTask::Failed(_) | AsyncTask::Cancelled => Ok(false),
-        }
-    }
-
-    pub(crate) fn cancel_task(&mut self, task_key: crate::arenas::TaskKey) -> WalrusResult<bool> {
-        let mut visited = FxHashSet::default();
-        let cancelled = self.cancel_task_recursive_internal(task_key, &mut visited)?;
-        if cancelled {
-            self.suspended_tasks.remove(&task_key);
-            self.wake_task_waiters(task_key)?;
-        }
-        Ok(cancelled)
-    }
-
-    pub(crate) fn run_queued_tasks(&mut self, span: Span) -> WalrusResult<()> {
-        while let Some(task_key) = self.next_runnable_task()? {
-            self.run_pending_task_to_completion(task_key, span)?;
-            self.refresh_waiting_tasks()?;
-        }
-        Ok(())
-    }
-
-    pub(crate) fn task_status_string(
-        &self,
-        task_key: crate::arenas::TaskKey,
-    ) -> WalrusResult<&'static str> {
-        let task = self.get_heap().get_task(task_key)?;
-        Ok(match task {
-            AsyncTask::Pending { .. }
-            | AsyncTask::Sleep { .. }
-            | AsyncTask::Timeout { .. }
-            | AsyncTask::Gather { .. }
-            | AsyncTask::Race { .. }
-            | AsyncTask::AllSettled { .. }
-            | AsyncTask::Channel(_)
-            | AsyncTask::UserRecv { .. } => "pending",
-            AsyncTask::Ready(_) => "ready",
-            AsyncTask::Failed(_) => "failed",
-            AsyncTask::Cancelled => "cancelled",
-        })
-    }
-
-    pub(crate) fn task_is_cancelled(&self, task_key: crate::arenas::TaskKey) -> WalrusResult<bool> {
-        let task = self.get_heap().get_task(task_key)?;
-        Ok(matches!(task, AsyncTask::Cancelled))
-    }
-
-    fn complete_task_on_frame_return(
-        &mut self,
-        task_key: Option<crate::arenas::TaskKey>,
-        result: Value,
-    ) -> WalrusResult<()> {
-        if let Some(task_key) = task_key {
-            let task = self.get_heap_mut().get_mut_task(task_key)?;
-            if matches!(task, AsyncTask::Pending { .. }) {
-                *task = AsyncTask::Ready(result);
-            }
-        }
-        Ok(())
-    }
-
-    fn set_task_failed(
-        &mut self,
-        task_key: crate::arenas::TaskKey,
-        failure: Value,
-    ) -> WalrusResult<()> {
-        *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Failed(failure);
-        self.wake_task_waiters(task_key)?;
-        Ok(())
-    }
-
-    fn task_failure_value_from_error(&mut self, err: WalrusError) -> Value {
-        let message = match err {
-            WalrusError::ThrownValue { message, .. } => message,
-            WalrusError::RuntimeErrorWithStackTrace { error, stack_trace } => {
-                format!("{error}\n{stack_trace}")
-            }
-            other => other.to_string(),
-        };
-        self.get_heap_mut().push(HeapValue::String(&message))
-    }
-
-    fn fail_task_with_error(
-        &mut self,
-        task_key: crate::arenas::TaskKey,
-        err: WalrusError,
-    ) -> WalrusResult<()> {
-        let failure = self.task_failure_value_from_error(err);
-        self.set_task_failed(task_key, failure)
-    }
-
-    fn enqueue_task(&mut self, task_key: crate::arenas::TaskKey) {
-        if !self.async_task_queue.contains(&task_key) {
-            self.async_task_queue.push_back(task_key);
-        }
-    }
-
-    fn wake_task_waiters(&mut self, task_key: crate::arenas::TaskKey) -> WalrusResult<()> {
-        let Some(waiters) = self.task_waiters.remove(&task_key) else {
-            return Ok(());
-        };
-
-        for waiter in waiters {
-            if self.suspended_tasks.contains_key(&waiter)
-                && !matches!(
-                    self.get_heap().get_task(waiter)?,
-                    AsyncTask::Ready(_) | AsyncTask::Failed(_) | AsyncTask::Cancelled
-                )
-            {
-                self.enqueue_task(waiter);
-            }
-        }
-
-        Ok(())
-    }
-
-    fn refresh_waiting_tasks(&mut self) -> WalrusResult<()> {
-        let mut watched = FxHashSet::default();
-        if let Some(suspended) = &self.suspended_main {
-            watched.insert(suspended.waiting_on);
-        }
-        watched.extend(self.task_waiters.keys().copied());
-
-        for task_key in watched {
-            if matches!(
-                self.poll_task_resolution(task_key)?,
-                TaskResolution::Ready(_) | TaskResolution::Failed(_) | TaskResolution::Cancelled
-            ) {
-                self.wake_task_waiters(task_key)?;
-            }
-        }
-
-        Ok(())
-    }
-
-    fn is_task_runnable(&mut self, task_key: crate::arenas::TaskKey) -> WalrusResult<bool> {
-        let task = self.get_heap().get_task(task_key)?;
-        if !matches!(task, AsyncTask::Pending { .. }) {
-            return Ok(false);
-        }
-
-        if let Some(suspended) = self.suspended_tasks.get(&task_key) {
-            return Ok(matches!(
-                self.poll_task_resolution(suspended.waiting_on)?,
-                TaskResolution::Ready(_) | TaskResolution::Failed(_) | TaskResolution::Cancelled
-            ));
-        }
-
-        Ok(true)
-    }
-
-    fn poll_task_resolution(
-        &mut self,
-        task_key: crate::arenas::TaskKey,
-    ) -> WalrusResult<TaskResolution> {
-        let task = self.get_heap().get_task(task_key)?.clone();
-        match task {
-            AsyncTask::Pending { .. } => Ok(TaskResolution::Pending),
-            AsyncTask::Ready(value) => Ok(TaskResolution::Ready(value)),
-            AsyncTask::Failed(value) => Ok(TaskResolution::Failed(value)),
-            AsyncTask::Cancelled => Ok(TaskResolution::Cancelled),
-            AsyncTask::Sleep { wake_at } => {
-                if Instant::now() >= wake_at {
-                    *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Ready(Value::Void);
-                    Ok(TaskResolution::Ready(Value::Void))
-                } else {
-                    Ok(TaskResolution::Pending)
-                }
-            }
-            AsyncTask::Timeout { task, deadline } => match self.poll_task_resolution(task)? {
-                TaskResolution::Ready(value) => {
-                    *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Ready(value);
-                    Ok(TaskResolution::Ready(value))
-                }
-                TaskResolution::Failed(value) => {
-                    *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Failed(value);
-                    Ok(TaskResolution::Failed(value))
-                }
-                TaskResolution::Cancelled => {
-                    *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Cancelled;
-                    Ok(TaskResolution::Cancelled)
-                }
-                TaskResolution::Pending => {
-                    if Instant::now() >= deadline {
-                        let message = self
-                            .get_heap_mut()
-                            .push(HeapValue::String("task timed out"));
-                        *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Failed(message);
-                        Ok(TaskResolution::Failed(message))
-                    } else {
-                        Ok(TaskResolution::Pending)
-                    }
-                }
-            },
-            AsyncTask::Channel(ref channel) => match channel.try_recv() {
-                Ok(Ok(io_result)) => {
-                    let value = self.io_result_to_value(io_result)?;
-                    *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Ready(value);
-                    Ok(TaskResolution::Ready(value))
-                }
-                Ok(Err(error_msg)) => {
-                    let error = self.get_heap_mut().push(HeapValue::String(&error_msg));
-                    *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Failed(error);
-                    Ok(TaskResolution::Failed(error))
-                }
-                Err(std::sync::mpsc::TryRecvError::Empty) => Ok(TaskResolution::Pending),
-                Err(std::sync::mpsc::TryRecvError::Disconnected) => {
-                    let error = self
-                        .get_heap_mut()
-                        .push(HeapValue::String("I/O worker thread disconnected"));
-                    *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Failed(error);
-                    Ok(TaskResolution::Failed(error))
-                }
-            },
-            AsyncTask::Gather { tasks } => {
-                let mut values = Vec::with_capacity(tasks.len());
-                for child in tasks {
-                    match self.poll_task_resolution(child)? {
-                        TaskResolution::Ready(value) => values.push(value),
-                        TaskResolution::Failed(value) => {
-                            *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Failed(value);
-                            return Ok(TaskResolution::Failed(value));
-                        }
-                        TaskResolution::Cancelled => {
-                            *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Cancelled;
-                            return Ok(TaskResolution::Cancelled);
-                        }
-                        TaskResolution::Pending => {
-                            return Ok(TaskResolution::Pending);
-                        }
-                    }
-                }
-
-                let list = self.get_heap_mut().push(HeapValue::List(values));
-                *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Ready(list);
-                Ok(TaskResolution::Ready(list))
-            }
-            AsyncTask::Race { tasks } => {
-                let mut all_failed = true;
-                let mut last_failure = None;
-                for child in tasks {
-                    match self.poll_task_resolution(child)? {
-                        TaskResolution::Ready(value) => {
-                            *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Ready(value);
-                            return Ok(TaskResolution::Ready(value));
-                        }
-                        TaskResolution::Failed(value) => {
-                            last_failure = Some(value);
-                        }
-                        TaskResolution::Cancelled => {}
-                        TaskResolution::Pending => {
-                            all_failed = false;
-                        }
-                    }
-                }
-                if all_failed {
-                    if let Some(failure) = last_failure {
-                        *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Failed(failure);
-                        Ok(TaskResolution::Failed(failure))
-                    } else {
-                        *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Cancelled;
-                        Ok(TaskResolution::Cancelled)
-                    }
-                } else {
-                    Ok(TaskResolution::Pending)
-                }
-            }
-            AsyncTask::AllSettled { tasks } => {
-                let mut results = Vec::with_capacity(tasks.len());
-                for child in tasks {
-                    match self.poll_task_resolution(child)? {
-                        TaskResolution::Ready(value) => {
-                            let status_str = self.get_heap_mut().push(HeapValue::String("ok"));
-                            let value_str = self.get_heap_mut().push(HeapValue::String("value"));
-                            let status_key = self.get_heap_mut().push(HeapValue::String("status"));
-                            let mut dict = FxHashMap::default();
-                            dict.insert(status_key, status_str);
-                            dict.insert(value_str, value);
-                            results.push(self.get_heap_mut().push(HeapValue::Dict(dict)));
-                        }
-                        TaskResolution::Failed(value) => {
-                            let status_str = self.get_heap_mut().push(HeapValue::String("error"));
-                            let error_str = self.get_heap_mut().push(HeapValue::String("error"));
-                            let status_key = self.get_heap_mut().push(HeapValue::String("status"));
-                            let mut dict = FxHashMap::default();
-                            dict.insert(status_key, status_str);
-                            dict.insert(error_str, value);
-                            results.push(self.get_heap_mut().push(HeapValue::Dict(dict)));
-                        }
-                        TaskResolution::Cancelled => {
-                            let status_str =
-                                self.get_heap_mut().push(HeapValue::String("cancelled"));
-                            let status_key = self.get_heap_mut().push(HeapValue::String("status"));
-                            let mut dict = FxHashMap::default();
-                            dict.insert(status_key, status_str);
-                            results.push(self.get_heap_mut().push(HeapValue::Dict(dict)));
-                        }
-                        TaskResolution::Pending => {
-                            return Ok(TaskResolution::Pending);
-                        }
-                    }
-                }
-                let list = self.get_heap_mut().push(HeapValue::List(results));
-                *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Ready(list);
-                Ok(TaskResolution::Ready(list))
-            }
-            AsyncTask::UserRecv { channel_id } => {
-                if channel_id < self.user_channels.len() {
-                    let mut buf = self.user_channels[channel_id].buffer.borrow_mut();
-                    if let Some(value) = buf.pop_front() {
-                        drop(buf);
-                        *self.get_heap_mut().get_mut_task(task_key)? = AsyncTask::Ready(value);
-                        return Ok(TaskResolution::Ready(value));
-                    }
-                    drop(buf);
-                    // Check if channel is closed with empty buffer
-                    if *self.user_channels[channel_id].closed.borrow() {
-                        *self.get_heap_mut().get_mut_task(task_key)? =
-                            AsyncTask::Ready(Value::Void);
-                        return Ok(TaskResolution::Ready(Value::Void));
-                    }
-                }
-                Ok(TaskResolution::Pending)
-            }
-        }
-    }
-
-    fn next_deadline_for_task(
-        &self,
-        task_key: crate::arenas::TaskKey,
-        visited: &mut FxHashSet<crate::arenas::TaskKey>,
-    ) -> WalrusResult<Option<Instant>> {
-        if !visited.insert(task_key) {
-            return Ok(None);
-        }
-
-        let task = self.get_heap().get_task(task_key)?;
-        match task {
-            AsyncTask::Sleep { wake_at } => Ok(Some(*wake_at)),
-            AsyncTask::Timeout { task, deadline } => {
-                let nested = self.next_deadline_for_task(*task, visited)?;
-                Ok(match nested {
-                    Some(value) => Some((*deadline).min(value)),
-                    None => Some(*deadline),
-                })
-            }
-            AsyncTask::Gather { tasks }
-            | AsyncTask::Race { tasks }
-            | AsyncTask::AllSettled { tasks } => {
-                let mut soonest: Option<Instant> = None;
-                for &child in tasks {
-                    if let Some(deadline) = self.next_deadline_for_task(child, visited)? {
-                        soonest = Some(match soonest {
-                            Some(current) => current.min(deadline),
-                            None => deadline,
-                        });
-                    }
-                }
-                Ok(soonest)
-            }
-            AsyncTask::Pending { .. }
-            | AsyncTask::Channel(_)
-            | AsyncTask::UserRecv { .. }
-            | AsyncTask::Ready(_)
-            | AsyncTask::Failed(_)
-            | AsyncTask::Cancelled => Ok(None),
-        }
-    }
-
-    fn next_scheduler_deadline(&self) -> WalrusResult<Option<Instant>> {
-        let mut watched = FxHashSet::default();
-        let mut deadline: Option<Instant> = None;
-
-        if let Some(suspended) = &self.suspended_main {
-            watched.insert(suspended.waiting_on);
-        }
-        watched.extend(self.task_waiters.keys().copied());
-
-        for task_key in watched {
-            let mut visited = FxHashSet::default();
-            if let Some(next) = self.next_deadline_for_task(task_key, &mut visited)? {
-                deadline = Some(match deadline {
-                    Some(current) => current.min(next),
-                    None => next,
-                });
-            }
-        }
-
-        Ok(deadline)
-    }
-
-    fn scheduler_has_pending_io(&self) -> WalrusResult<bool> {
-        let mut watched = FxHashSet::default();
-        if let Some(suspended) = &self.suspended_main {
-            watched.insert(suspended.waiting_on);
-        }
-        watched.extend(self.task_waiters.keys().copied());
-
-        for task_key in watched {
-            let mut visited = FxHashSet::default();
-            if self.task_has_pending_io(task_key, &mut visited)? {
-                return Ok(true);
-            }
-        }
-
-        Ok(false)
-    }
-
-    fn wait_for_scheduler_progress(&mut self) -> WalrusResult<()> {
-        let deadline = self.next_scheduler_deadline()?;
-        let has_io = self.scheduler_has_pending_io()?;
-
-        if deadline.is_none() && !has_io {
-            return Err(WalrusError::GenericError {
-                message:
-                    "Event loop deadlock: all suspended work is waiting with no runnable tasks"
-                        .to_string(),
-            });
-        }
-
-        if let Some(deadline) = deadline {
-            let now = Instant::now();
-            if deadline > now {
-                let timeout = deadline.duration_since(now);
-                let _ = self.io_wakeup_rx.recv_timeout(timeout);
-            }
-        } else {
-            let _ = self.io_wakeup_rx.recv();
-        }
-
-        Ok(())
-    }
-
-    fn main_waiting_task_resolved(&mut self) -> WalrusResult<bool> {
-        let Some(suspended) = &self.suspended_main else {
-            return Ok(false);
-        };
-
-        Ok(matches!(
-            self.poll_task_resolution(suspended.waiting_on)?,
-            TaskResolution::Ready(_) | TaskResolution::Failed(_) | TaskResolution::Cancelled
-        ))
-    }
-
-    fn resume_main_if_ready(&mut self) -> WalrusResult<bool> {
-        if !self.main_waiting_task_resolved()? {
-            return Ok(false);
-        }
-
-        let suspended = self
-            .suspended_main
-            .take()
-            .expect("main suspension should exist when ready");
-        self.restore_context(suspended.context);
-        Ok(true)
-    }
-
-    fn next_runnable_task(&mut self) -> WalrusResult<Option<crate::arenas::TaskKey>> {
-        while let Some(task_key) = self.async_task_queue.pop_front() {
-            if self.is_task_runnable(task_key)? {
-                return Ok(Some(task_key));
-            }
-        }
-        Ok(None)
-    }
-
-    fn run_pending_task_to_completion(
-        &mut self,
-        task_key: crate::arenas::TaskKey,
-        span: Span,
-    ) -> WalrusResult<()> {
-        let task_snapshot = self.get_heap().get_task(task_key)?.clone();
-        let AsyncTask::Pending { function, args } = task_snapshot else {
-            return Ok(());
-        };
-
-        let caller_context = self.take_context();
-        let suspended_task = self.suspended_tasks.remove(&task_key);
-        let function = self.get_heap().get_function(function)?.clone();
-        match function {
-            WalrusFunction::Vm(func) => {
-                if args.len() != func.arity {
-                    self.restore_context(caller_context);
-                    return self.fail_task_with_error(
-                        task_key,
-                        WalrusError::InvalidArgCount {
-                            name: func.name.clone(),
-                            expected: func.arity,
-                            got: args.len(),
-                            span,
-                            src: self.source_ref.source().into(),
-                            filename: self.source_ref.filename().into(),
-                        },
-                    );
-                }
-
-                let context = if let Some(suspended) = suspended_task {
-                    suspended.context
-                } else {
-                    ExecutionContext {
-                        stack: Vec::new(),
-                        locals: args,
-                        call_stack: vec![CallFrame {
-                            return_ip: 0,
-                            frame_pointer: 0,
-                            stack_pointer: 0,
-                            instructions: Rc::clone(&func.code),
-                            function_name: func.name.clone(),
-                            return_override: None,
-                            module_binding: func.module_binding.clone(),
-                            awaiting_task: Some(task_key),
-                        }],
-                        exception_handlers: Vec::new(),
-                        ip: 0,
-                    }
-                };
-
-                self.restore_context(context);
-                let outcome = self.run_inner();
-                let task_context = self.take_context();
-                self.restore_context(caller_context);
-
-                match outcome {
-                    Ok(RunSignal::Returned(_)) => {
-                        if matches!(
-                            self.poll_task_resolution(task_key)?,
-                            TaskResolution::Pending
-                        ) {
-                            let failure = self.get_heap_mut().push(HeapValue::String(
-                                "Async task exited without returning a result",
-                            ));
-                            self.set_task_failed(task_key, failure)?;
-                        } else {
-                            self.wake_task_waiters(task_key)?;
-                        }
-                    }
-                    Ok(RunSignal::Suspended(waiting_on)) => {
-                        self.suspended_tasks.insert(
-                            task_key,
-                            SuspendedExecution {
-                                context: task_context,
-                                waiting_on,
-                            },
-                        );
-                        self.task_waiters
-                            .entry(waiting_on)
-                            .or_default()
-                            .push(task_key);
-                    }
-                    Err(err) => {
-                        let failure = self.task_failure_value_from_error(err);
-                        self.set_task_failed(task_key, failure)?;
-                    }
-                }
-            }
-            WalrusFunction::Native(native_fn) => {
-                let result = match self.call_native(native_fn, args, span) {
-                    Ok(value) => value,
-                    Err(err) => {
-                        self.restore_context(caller_context);
-                        return self.fail_task_with_error(task_key, err);
-                    }
-                };
-                self.complete_task_on_frame_return(Some(task_key), result)?;
-                self.restore_context(caller_context);
-                self.wake_task_waiters(task_key)?;
-            }
-        }
-
-        Ok(())
-    }
-
-    fn call_exported_function(
-        &mut self,
-        function: WalrusFunction,
-        args: Vec<Value>,
-        span: Span,
-    ) -> WalrusResult<Option<Value>> {
-        match function {
-            WalrusFunction::Native(native) => {
-                let result = self.call_native(native, args, span)?;
-                Ok(Some(result))
-            }
-            WalrusFunction::Vm(func) => {
-                if args.len() != func.arity {
-                    return Err(WalrusError::InvalidArgCount {
-                        name: func.name.clone(),
-                        expected: func.arity,
-                        got: args.len(),
-                        span,
-                        src: self.source_ref.source().into(),
-                        filename: self.source_ref.filename().into(),
-                    });
-                }
-
-                if func.is_async {
-                    let task = self
-                        .create_task_for_cloned_function(WalrusFunction::Vm(func.clone()), args);
-                    return Ok(Some(task));
-                }
-
-                let new_frame = CallFrame {
-                    return_ip: self.ip,
-                    frame_pointer: self.locals.len(),
-                    stack_pointer: self.stack.len(),
-                    instructions: Rc::clone(&func.code),
-                    function_name: func.name.clone(),
-                    return_override: None,
-                    module_binding: func.module_binding.clone(),
-                    awaiting_task: None,
-                };
-
-                self.call_stack.push(new_frame);
-                self.locals.extend(args);
-                self.ip = 0;
-                Ok(None)
-            }
-        }
-    }
-
-    /// Helper to extract string from Value
-    pub(crate) fn value_to_string(&self, value: Value, span: Span) -> WalrusResult<String> {
-        match value {
-            Value::String(key) => Ok(self.get_heap().get_string(key)?.to_string()),
-            _ => Err(WalrusError::TypeMismatch {
-                expected: "string".to_string(),
-                found: value.get_type().to_string(),
-                span,
-                src: self.source_ref.source().into(),
-                filename: self.source_ref.filename().into(),
-            }),
-        }
-    }
-
-    /// Helper to extract int from Value
-    pub(crate) fn value_to_int(&self, value: Value, span: Span) -> WalrusResult<i64> {
-        match value {
-            Value::Int(n) => Ok(n),
-            _ => Err(WalrusError::TypeMismatch {
-                expected: "int".to_string(),
-                found: value.get_type().to_string(),
-                span,
-                src: self.source_ref.source().into(),
-                filename: self.source_ref.filename().into(),
-            }),
-        }
-    }
-
-    /// Helper to extract numeric values (int or float) as f64
-    pub(crate) fn value_to_number(&self, value: Value, span: Span) -> WalrusResult<f64> {
-        match value {
-            Value::Int(n) => Ok(n as f64),
-            Value::Float(FloatOrd(f)) => Ok(f),
-            _ => Err(WalrusError::TypeMismatch {
-                expected: "number".to_string(),
-                found: value.get_type().to_string(),
-                span,
-                src: self.source_ref.source().into(),
-                filename: self.source_ref.filename().into(),
-            }),
-        }
-    }
-
-    /// Normalize an index (supports negative indices) against a character length.
-    fn normalize_index(index: i64, len: usize) -> Option<usize> {
-        let len = len as i64;
-        let normalized = if index < 0 { index + len } else { index };
-        if normalized < 0 || normalized >= len {
-            None
-        } else {
-            Some(normalized as usize)
-        }
-    }
-
-    /// Convert a character index to a byte offset.
-    /// Returns `s.len()` when `char_index` is exactly one-past-the-end.
-    fn char_to_byte_offset(s: &str, char_index: usize) -> Option<usize> {
-        if char_index == 0 {
-            return Some(0);
-        }
-        if char_index == s.chars().count() {
-            return Some(s.len());
-        }
-        s.char_indices().nth(char_index).map(|(offset, _)| offset)
-    }
-
-    /// Remove handlers that are no longer reachable from the current execution point.
-    ///
-    /// Handlers become stale when:
-    /// - Their frame has been popped.
-    /// - Control flow in the same frame has jumped out of the protected try range.
-    fn prune_exception_handlers(&mut self) {
-        if self.exception_handlers.is_empty() {
-            return;
-        }
-
-        let current_frame = self.current_frame_index();
-        while let Some(handler) = self.exception_handlers.last().copied() {
-            let stale_frame = handler.frame_index > current_frame;
-            let out_of_range_in_frame = handler.frame_index == current_frame
-                && (self.ip < handler.start_ip || self.ip >= handler.end_ip);
-
-            if stale_frame || out_of_range_in_frame {
-                self.exception_handlers.pop();
-            } else {
-                break;
-            }
-        }
-    }
-
-    /// Clear all handlers that belong to `frame_index` or deeper.
-    /// Used when a frame is reused (tail call) or dropped.
-    fn clear_exception_handlers_from_frame(&mut self, frame_index: usize) {
-        while let Some(handler) = self.exception_handlers.last().copied() {
-            if handler.frame_index >= frame_index {
-                self.exception_handlers.pop();
-            } else {
-                break;
-            }
-        }
-    }
-
-    /// Raise a thrown value, transferring control to the nearest active catch handler.
-    fn throw_value(&mut self, value: Value, span: Span) -> WalrusResult<()> {
-        while let Some(handler) = self.exception_handlers.last().copied() {
-            if handler.frame_index > self.current_frame_index() {
-                self.exception_handlers.pop();
-                continue;
-            }
-
-            self.exception_handlers.pop();
-
-            // Unwind call frames until we reach the frame that owns this handler.
-            while self.current_frame_index() > handler.frame_index {
-                let frame = self
-                    .call_stack
-                    .pop()
-                    .expect("Call stack should never be empty while unwinding");
-                self.locals.truncate(frame.frame_pointer);
-                self.stack.truncate(frame.stack_pointer);
-            }
-
-            self.locals.truncate(handler.locals_len);
-            self.stack.truncate(handler.stack_len);
-            self.ip = handler.catch_ip;
-            self.push(value);
-            return Ok(());
-        }
-
-        let message = self.stringify_value(value)?;
-        Err(WalrusError::ThrownValue {
-            message,
-            span,
-            src: self.source_ref.source().into(),
-            filename: self.source_ref.filename().into(),
-        })
-    }
-
     /// Run the VM and add stack trace information to any errors
     pub fn run(&mut self) -> WalrusResult<Value> {
         loop {
@@ -4429,87 +2756,5 @@ impl<'a> VM<'a> {
                 }
             }
         }
-    }
-
-    #[inline(always)]
-    fn push(&mut self, value: Value) {
-        self.stack.push(value);
-    }
-
-    #[inline]
-    fn pop(&mut self, op: Opcode, span: Span) -> WalrusResult<Value> {
-        self.stack.pop().ok_or_else(|| WalrusError::StackUnderflow {
-            op,
-            span,
-            src: self.source_ref.source().to_string(),
-            filename: self.source_ref.filename().to_string(),
-        })
-    }
-
-    /// Fast path pop - only use when stack is guaranteed to have values
-    #[inline(always)]
-    fn pop_unchecked(&mut self) -> Value {
-        // SAFETY: caller guarantees stack is not empty
-        unsafe { self.stack.pop().unwrap_unchecked() }
-    }
-
-    fn pop_n(&mut self, n: usize, op: Opcode, span: Span) -> WalrusResult<Vec<Value>> {
-        if self.stack.len() < n {
-            return Err(WalrusError::StackUnderflow {
-                op,
-                span,
-                src: self.source_ref.source().to_string(),
-                filename: self.source_ref.filename().to_string(),
-            });
-        }
-
-        let split_at = self.stack.len() - n;
-        Ok(self.stack.split_off(split_at))
-    }
-
-    fn stack_trace(&self) {
-        for (i, frame) in self.stack.iter().enumerate() {
-            debug!("| {} | {}: {}", self.function_name(), i, frame);
-        }
-    }
-
-    /// Run the debugger prompt and return the command
-    fn run_debugger_prompt(
-        &mut self,
-        instructions: &InstructionSet,
-    ) -> WalrusResult<debugger::DebuggerCommand> {
-        // Build call stack info for debugger
-        let call_stack: Vec<debugger::DebugCallFrame> = self
-            .call_stack
-            .iter()
-            .map(|f| debugger::DebugCallFrame {
-                function_name: if f.function_name.is_empty() {
-                    "<fn>".to_string()
-                } else {
-                    f.function_name.clone()
-                },
-                return_ip: f.return_ip,
-                frame_pointer: f.frame_pointer,
-            })
-            .collect();
-
-        let ctx = debugger::DebugContext {
-            ip: self.ip,
-            stack: &self.stack,
-            locals: &self.locals,
-            globals: &self.globals,
-            call_stack: &call_stack,
-            debug_info: instructions.debug_info.as_ref(),
-            instructions: &instructions.instructions,
-            source: self.source_ref.source(),
-        };
-
-        let cmd = if let Some(ref mut dbg) = self.debugger {
-            debugger::debug_prompt(dbg, &ctx)
-        } else {
-            debugger::DebuggerCommand::Continue
-        };
-
-        Ok(cmd)
     }
 }
